work on board , test -v -h passed

+++ srsLTE/lib/include/srslte/common/bounded_bitset.h
@@ -52,7 +52,7 @@ public:
   void resize(size_t new_size) noexcept
   {
     if (new_size > max_size()) {
-      printf("ERROR: bitset resize out of bounds: %lu>=%lu\n", max_size(), new_size);
+      printf("ERROR: bitset resize out of bounds: %lu>=%lu\n", (long unsigned int)max_size(), (long unsigned int)new_size);
       return;
     } else if (new_size == cur_size) {
       return;
@@ -67,7 +67,7 @@ public:
   void set(size_t pos) noexcept
   {
     if (pos >= size()) {
-      printf("ERROR: bitset out of bounds: %lu>=%lu\n", pos, size());
+      printf("ERROR: bitset out of bounds: %lu>=%lu\n", (long unsigned int)pos, (long unsigned int)size());
       return;
     }
     set_(pos);
@@ -76,7 +76,7 @@ public:
   void reset(size_t pos) noexcept
   {
     if (pos >= size()) {
-      printf("ERROR: bitset out of bounds: %lu>=%lu\n", pos, size());
+      printf("ERROR: bitset out of bounds: %lu>=%lu\n", (long unsigned int)pos, (long unsigned int)size());
       return;
     }
     reset_(pos);
@@ -92,7 +92,7 @@ public:
   bool test(size_t pos) const noexcept
   {
     if (pos >= size()) {
-      printf("ERROR: bitset out of bounds: %lu>=%lu\n", pos, size());
+      printf("ERROR: bitset out of bounds: %lu>=%lu\n", (long unsigned int)pos, (long unsigned int)size());
       return false;
     }
     return test_(pos);
@@ -110,7 +110,7 @@ public:
   bounded_bitset<N, reversed>& fill(size_t startpos, size_t endpos, bool value = true) noexcept
   {
     if (endpos > size() or startpos > endpos) {
-      printf("ERROR: bounds (%lu, %lu) are not valid for bitset of size: %lu\n", startpos, endpos, size());
+      printf("ERROR: bounds (%lu, %lu) are not valid for bitset of size: %lu\n", (long unsigned int)startpos, (long unsigned int)endpos, (long unsigned int)size());
       return *this;
     }
     // NOTE: can be optimized
@@ -154,7 +154,7 @@ public:
   bool any(size_t start, size_t stop) const noexcept
   {
     if (start > stop or stop > size()) {
-      printf("ERROR: bounds (%lu, %lu) are not valid for bitset of size: %lu\n", start, stop, size());
+      printf("ERROR: bounds (%lu, %lu) are not valid for bitset of size: %lu\n", (long unsigned int)start, (long unsigned int)stop, (long unsigned int)size());
       return false;
     }
     // NOTE: can be optimized
@@ -198,7 +198,7 @@ public:
   bounded_bitset<N, reversed>& operator|=(const bounded_bitset<N, reversed>& other) noexcept
   {
     if (other.size() != size()) {
-      printf("ERROR: operator|= called for bitsets of different sizes (%lu!=%lu)\n", size(), other.size());
+      printf("ERROR: operator|= called for bitsets of different sizes (%lu!=%lu)\n", (long unsigned int)size(), (long unsigned int)other.size());
       return *this;
     }
     for (size_t i = 0; i < nof_words_(); ++i) {
@@ -210,7 +210,7 @@ public:
   bounded_bitset<N, reversed>& operator&=(const bounded_bitset<N, reversed>& other) noexcept
   {
     if (other.size() != size()) {
-      printf("ERROR: operator&= called for bitsets of different sizes (%lu!=%lu)\n", size(), other.size());
+      printf("ERROR: operator&= called for bitsets of different sizes (%lu!=%lu)\n", (long unsigned int)size(), (long unsigned int)other.size());
       return *this;
     }
     for (size_t i = 0; i < nof_words_(); ++i) {
@@ -249,7 +249,7 @@ public:
   uint64_t to_uint64() const noexcept
   {
     if (nof_words_() > 1) {
-      printf("ERROR: cannot convert bitset of size %lu bits to uint64_t\n", size());
+      printf("ERROR: cannot convert bitset of size %lu bits to uint64_t\n", (long unsigned int)size());
       return 0;
     }
     return get_word_(0);
@@ -262,7 +262,7 @@ public:
     size_t count = 0;
 
     for (int i = nof_words_() - 1; i >= 0; --i) {
-      count += sprintf(&cstr[count], "%016lx", buffer[i]);
+      count += sprintf(&cstr[count], "%016lx", (long unsigned int)buffer[i]);
     }
 
     size_t skip = nof_words_() * bits_per_word / 4 - nof_digits;
